services/web/backend/app/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
pyotp==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
